myrequirements.txt
django-crispy-forms==1.9.2
django-cors-headers=3.7.0
djangorestframework=3.12.2
#sudo apt-get install tree
